alert-bridge/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.4.2
pyyaml==6.0.1
python-json-logger==2.0.7
prometheus-client==0.17.1 